+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 #### 🚀 Updates
 
+- Added support for `#[setting(nested = NestedConfig)]` on fields, where the nested config name can
+  be explicitly defined if we fail to detect it. This is useful for extremely complex/composed
+  types.
 - Added support for explicit Serde deserialize and serialize renaming on containers and fields:
   `#[serde(rename(deserialize = "de_name", serialize = "ser_name"))]`.
 - Updated Serde `alias` to support multiple aliases: `#[serde(alias = "alias1", alias = "alias2")]`